[Android]fix slider border radius invalid problem (#2658)
Demo: https://alibaba.github.io/rax/jsplayground/60e09c4e-9ca5-4eb7-b6a0-f38040e0c626
